+++ joulescope_ui/widgets/__init__.py
@@ -12,10 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .accumulator import AccumulatorWidget
-from .device_control import DeviceControlWidget
 from .double_slider import DoubleSlider
 from .draggable_list_widget import DraggableListWidget
+
+from .accumulator import AccumulatorWidget
+from .device_control import DeviceControlWidget
 # from .example import ExampleWidget
 from .memory import MemoryWidget
 from .hamburger import HamburgerWidget